+++ src/init.c
@@ -4,13 +4,13 @@
 #include <Rinternals.h>
 
 extern SEXP read_png_(SEXP src_, SEXP type_, SEXP flags_,SEXP rgba_, SEXP avoid_transpose_);
-extern SEXP write_png_(SEXP image_ , SEXP file_, SEXP palette_, SEXP use_filter_, SEXP compression_level_, SEXP avoid_transpose_, SEXP bits_);
+extern SEXP write_png_(SEXP image_ , SEXP file_, SEXP palette_, SEXP use_filter_, SEXP compression_level_, SEXP avoid_transpose_, SEXP bits_, SEXP trns_);
 extern SEXP get_png_info_(SEXP src_);
 
 
 static const R_CallMethodDef CEntries[] = {
   {"read_png_"   , (DL_FUNC) &read_png_ , 5},
-  {"write_png_"  , (DL_FUNC) &write_png_, 7},
+  {"write_png_"  , (DL_FUNC) &write_png_, 8},
 
   {"get_png_info_", (DL_FUNC) &get_png_info_, 1},
   {NULL , NULL, 0}